Add mobile integration tests
To emulate the file upload, we record the command with real build files
but use fake binary files for testing. This allows us not to commit
the read files to this repository which makes its size quite large.

By separating one command test per test file, we can control
which test should be run on which platform.