Normalize command line for integration test
`autify-with-proxy` naively relies on the passed command line to
determine which polly recording file should be used.
This is not robust for the consumers like GitHub Actions
because they need to pass the exactly same command line.

This commit introduces `normalizeCommand()` function so that
the consumers can pass more flexible command such as
`https://app.autify.com/projects/0000/scenarios/0000` and
it recognizes all known input patterns and replaces them to
the real and recordable values as well as sorts, concatenates
flag and value by `=`, etc.