Update Autify Connect Client and revert fake path
Simply updating Autify Connect Client to 0.6.12 but we realized
the previous commit confuses local recording because fake binary
is installed to the same path of the real one.

This commit also fixes that issue and add useful commands for
recording. Now, recording uses the real binary and test uses the fake.